Remove browser script `window.onerror` handler
The window handler is a nice idea since it helps prevent Phantom from
hanging in the case where there's an uncaught exception before the test
harness is set up to catch errors (see examples/fatal-errors for an
example case). That's an edge case, though, and it prevents Phantom
environments from operating as a cleanroom environment. If a script
ends up overwriting window.onerror, our tests hang. IMO this is a more
likey case, and was very frustrating to debug.

The proper fix for this is to load a PhantomJS script, create a page
using require('page'), and set an onError handler on that page, then
load the URL inside it. This gives you a cleanroom environment and
prevents Phantom from hanging in the event that there's an uncaught
exception.